llama3/requirements.txt
fastapi
llama-cpp-python
uvicorn
huggingface-hub